+++ library/SolverUtils/Filters/FilterHistoryPoints.cpp
@@ -39,6 +39,8 @@
 
 #include <boost/format.hpp>
 
+using namespace std;
+
 namespace Nektar
 {
 namespace SolverUtils