+++ library/Timings/TimingHDGHelmSolve2D.cpp
@@ -6,6 +6,7 @@
 #include <SpatialDomains/MeshGraph2D.h>
 #include <MultiRegions/DisContField2D.h>
 
+using namespace std;
 using namespace Nektar;
 
 std::string PortablePath(const boost::filesystem::path& path);